wrote helpers function for e2e test, working in actors and films tests so far
