+++ lib/app.js
@@ -4,11 +4,13 @@ const app = express();
 const morgan = require('morgan');
 const { handler } = require('./util/errors');
 
-app.use(morgan('dev', {
-    skip() {
-        return process.env.NODE_ENV === 'test';
-    }
-}));
+// app.use(morgan('dev', {
+//     skip() {
+//         return process.env.NODE_ENV === 'test';
+//     }
+// }));
+
+app.use(morgan('dev'));
 
 app.use(express.json());
