+++ test/e2e/actors.test.js
@@ -1,80 +1,9 @@
-const { dropCollection } = require('./db');
 const request = require('supertest');
 const app = require('../../lib/app');
+const { getActors, getFilms } = require('./helpers');
 
 
 describe('validates a vertical slice of the Actor route', () => {
-
-    beforeEach(() => {
-        return dropCollection('actors');
-    });
-    beforeEach(() => {
-        return dropCollection('films');
-    });
-
-    let createdActors;
-    let createdFilms;
-
-    let actors =  [{
-        name: 'Anna Peel',
-        dob: new Date(),
-        pob: 'Brazil'
-    },
-    {
-        name: 'Lud Orange',
-        dob: new Date(),
-        pob: 'Florida'
-    },
-    {
-        name: 'Ocado Pitt',
-        dob: new Date(),
-        pob: 'Mexico'
-    }
-    ];
-
-    let films =  [{
-        title: 'Bladecrawler',
-        released: 1991,
-        cast: [{
-            role: 'lead',
-        }]
-    },
-    {
-        title: 'Bladewalker',
-        released: 1992,
-        cast: [{
-            role: 'Deckard',
-        }]
-    }];
-        
-    const createActor = actor => {
-        return request(app)
-            .post('/api/actors')
-            .send(actor)
-            .then(res => res.body);
-    };
-
-    const createFilm = film => {
-        return request(app)
-            .post('/api/films')
-            .send(film)
-            .then(res => res.body);
-    };
-
-    beforeEach(() => {
-        return Promise.all(actors.map(createActor)).then(actorsRes => {
-            createdActors = actorsRes;
-            films[0].cast[0].actor = createdActors[0]._id;
-            films[1].cast[0].actor = createdActors[1]._id;
-        });
-    });
-        
-    beforeEach(() => {
-        return Promise.all(films.map(createFilm)).then(filmsRes => {
-            createdFilms = filmsRes;
-        });
-    });
-
 
     it('Posts to Actor', () => {
         return request(app)
@@ -96,16 +25,18 @@ describe('validates a vertical slice of the Actor route', () => {
     });
 
     it('gets all Actors', () => {
+        const createdActors = getActors();
         return request(app)
             .get('/api/actors')
             .then(res => {
                 expect(res.body).toContainEqual({ _id: createdActors[0]._id, name: createdActors[0].name });
                 expect(res.body).toContainEqual({ _id: createdActors[1]._id, name: createdActors[1].name });
-                expect(res.body).toContainEqual({ _id: createdActors[2]._id, name: createdActors[2].name });
             });
     });
 
     it('gets a actor by id', () => {
+        const createdActors = getActors();
+        const createdFilms = getFilms();
         return request(app)
             .get(`/api/actors/${createdActors[1]._id}`)
             .then(res => {
@@ -125,6 +56,7 @@ describe('validates a vertical slice of the Actor route', () => {
     });
 
     it('deletes a actor that is not in any film casts', () => {
+        const createdActors = getActors();
         return request(app)
             .delete(`/api/actors/${createdActors[2]._id}`)
             .then(modifiedList => {
@@ -133,6 +65,7 @@ describe('validates a vertical slice of the Actor route', () => {
     });
 
     it('responds with an error when asked to delete an actor referenced in a film', () => {
+        const createdActors = getActors();
         return request(app)
             .delete(`/api/actors/${createdActors[1]._id}`)
             .then(modifiedList => {
@@ -141,6 +74,7 @@ describe('validates a vertical slice of the Actor route', () => {
     });
 
     it('updates a actor by id', () => {
+        const createdActors = getActors();
         return request(app)
             .put(`/api/actors/${createdActors[1]._id}`)
             .send({
