+++ test/e2e/films.test.js
@@ -1,110 +1,13 @@
-const { dropCollection } = require('./db');
 const request = require('supertest');
 const app = require('../../lib/app');
-
+const { getActors, getFilms, getStudios } = require('./helpers');
 
 describe('validates a vertical slice of the Films route', () => {
-    beforeEach(() => {
-        return dropCollection('studios');
-    });
-    beforeEach(() => {
-        return dropCollection('actors');
-    });
-    beforeEach(() => {
-        return dropCollection('films');
-    });
-
-    let createdActors;
-    let createdStudios;
-    let createdFilms;
-
-    let actors =  [{
-        name: 'Anna Peel',
-        dob: new Date(),
-        pob: 'Brazil'
-    },
-    {
-        name: 'Lud Orange',
-        dob: new Date(),
-        pob: 'Florida'
-    }];
-    
-    let studios =  [{
-        name: 'Compost Cinema', 
-        address: {
-            city: 'Portland',
-            state: 'OR',
-            country: 'United States'
-        }
-    },
-    {
-        name: 'Compost Cinema2', 
-        address: {
-            city: 'Portland2',
-            state: 'OR',
-            country: 'United States'
-        }
-    }];
-
-    let films =  [{
-        title: 'Bladecrawler',
-        released: 1991,
-        cast: [{
-            role: 'lead',
-        }]
-    },
-    {
-        title: 'Bladewalker',
-        released: 1992,
-        cast: [{
-            role: 'Deckard',
-        }]
-    }];
-        
-    const createActor = actor => {
-        return request(app)
-            .post('/api/actors')
-            .send(actor)
-            .then(res => res.body);
-    };
-    
-    const createStudio = studio => {
-        return request(app)
-            .post('/api/studios')
-            .send(studio)
-            .then(res => res.body);
-    };
-
-    const createFilm = film => {
-        return request(app)
-            .post('/api/films')
-            .send(film)
-            .then(res => res.body);
-    };
-
-    beforeEach(() => {
-        return Promise.all(actors.map(createActor)).then(actorsRes => {
-            createdActors = actorsRes;
-            films[0].cast[0].actor = createdActors[0]._id;
-            films[1].cast[0].actor = createdActors[1]._id;
-        });
-    });
-
-    beforeEach(() => {
-        return Promise.all(studios.map(createStudio)).then(studiosRes => {
-            createdStudios = studiosRes;
-            films[0].studio = createdStudios[0]._id;
-            films[1].studio = createdStudios[1]._id;
-        });
-    });
-        
-    beforeEach(() => {
-        return Promise.all(films.map(createFilm)).then(filmsRes => {
-            createdFilms = filmsRes;
-        });
-    });
 
     it('Posts to Films', () => {
+        const createdStudios = getStudios();
+        const createdActors = getActors();
+
         return request(app)
             .post('/api/films')
             .send({
@@ -133,6 +36,9 @@ describe('validates a vertical slice of the Films route', () => {
     });
     
     it('gets all Films', () => {
+        const createdFilms = getFilms();
+        const createdStudios = getStudios();
+
         return request(app)
             .get('/api/films')
             .then(res => {
@@ -158,6 +64,10 @@ describe('validates a vertical slice of the Films route', () => {
     });
 
     it('gets a film by id', () => {
+        const createdFilms = getFilms();
+        const createdStudios = getStudios();
+        const createdActors = getActors();
+
         return request(app)
             .get(`/api/films/${createdFilms[1]._id}`)
             .then(res => {
@@ -182,6 +92,8 @@ describe('validates a vertical slice of the Films route', () => {
     });
 
     it('deletes a film by id', () => {
+        const createdFilms = getFilms();
+
         return request(app)
             .delete(`/api/films/${createdFilms[1]._id}`)
             .then(modifiedList => {
